# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface ClientProfile {
  userId: string;
  username: string;
  email: string;
  name: string;
  company: string;
}

export interface UpdateClientProfileInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /** @maxLength 160 */
  company: string;
}

export type ClientRequestKind = typeof ClientRequestKind[keyof typeof ClientRequestKind];


export const ClientRequestKind = {
  print: 'print',
} as const;

export type ClientRequestStatus = typeof ClientRequestStatus[keyof typeof ClientRequestStatus];


export const ClientRequestStatus = {
  submitted: 'submitted',
  reviewing: 'reviewing',
  quoted: 'quoted',
  scheduled: 'scheduled',
  completed: 'completed',
  suspended: 'suspended',
} as const;

export interface ClientRequest {
  id: string;
  reference: string;
  kind: ClientRequestKind;
  projectName: string;
  serviceSlug: string;
  status: ClientRequestStatus;
  statusAr: string;
  statusEn: string;
  material: string;
  finish: string;
  quantity: number;
  timeline: string;
  notes: string;
  fileName?: string;
  createdAt: string;
}

export interface ClientOverview {
  profile: ClientProfile;
  requests: ClientRequest[];
}

export interface CreateClientPrintRequestInput {
  /**
     * @minLength 2
     * @maxLength 160
     */
  projectName: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  material: string;
  /**
     * @minimum 1
     * @maximum 1000
     */
  quantity: number;
  /**
     * @minLength 1
     * @maxLength 80
     */
  finish: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  timeline: string;
  /** @maxLength 3000 */
  notes: string;
  /** @maxLength 255 */
  fileName?: string;
}

export type ClientConsultationKind = typeof ClientConsultationKind[keyof typeof ClientConsultationKind];


export const ClientConsultationKind = {
  consultation: 'consultation',
  specialist: 'specialist',
} as const;

export type ClientConsultationStatus = typeof ClientConsultationStatus[keyof typeof ClientConsultationStatus];


export const ClientConsultationStatus = {
  submitted: 'submitted',
  reviewing: 'reviewing',
  contacted: 'contacted',
  completed: 'completed',
  suspended: 'suspended',
} as const;

export type ClientConsultationProviderType = typeof ClientConsultationProviderType[keyof typeof ClientConsultationProviderType];


export const ClientConsultationProviderType = {
  person: 'person',
  company: 'company',
  guide: 'guide',
} as const;

export interface ClientConsultation {
  id: string;
  reference: string;
  kind: ClientConsultationKind;
  status: ClientConsultationStatus;
  statusAr: string;
  statusEn: string;
  title: string;
  details: string;
  specialty?: string;
  providerType?: ClientConsultationProviderType;
  preferredProvider?: string;
  createdAt: string;
}

export type ClientConsultationInputKind = typeof ClientConsultationInputKind[keyof typeof ClientConsultationInputKind];


export const ClientConsultationInputKind = {
  consultation: 'consultation',
  specialist: 'specialist',
} as const;

export type ClientConsultationInputProviderType = typeof ClientConsultationInputProviderType[keyof typeof ClientConsultationInputProviderType];


export const ClientConsultationInputProviderType = {
  person: 'person',
  company: 'company',
  guide: 'guide',
} as const;

export interface ClientConsultationInput {
  kind: ClientConsultationInputKind;
  /**
     * @minLength 2
     * @maxLength 160
     */
  title: string;
  /**
     * @minLength 10
     * @maxLength 4000
     */
  details: string;
  /** @maxLength 120 */
  specialty?: string;
  providerType?: ClientConsultationInputProviderType;
  /** @maxLength 160 */
  preferredProvider?: string;
}

export interface ServiceSummary {
  slug: string;
  titleAr: string;
  titleEn: string;
  descriptionAr: string;
  descriptionEn: string;
  category: string;
  duration: string;
  accent: string;
}

export interface ServiceGalleryItem {
  image: string;
  titleAr: string;
  titleEn: string;
  descriptionAr: string;
  descriptionEn: string;
}

export type ServiceDetail = ServiceSummary & {
  highlightsAr: string[];
  highlightsEn: string[];
  workflowAr: string[];
  workflowEn: string[];
  gallery: ServiceGalleryItem[];
};

export interface Material {
  id: string;
  name: string;
  applicationAr: string;
  applicationEn: string;
  characteristicsAr: string;
  characteristicsEn: string;
  examplesAr: string;
  examplesEn: string;
  pricePerGram: number;
  leadDays: number;
}

export interface Stat {
  value: string;
  labelAr: string;
  labelEn: string;
}

export interface DesignStage {
  number: number;
  titleAr: string;
  titleEn: string;
  descriptionAr: string;
  descriptionEn: string;
}

export interface Differentiator {
  titleAr: string;
  titleEn: string;
  descriptionAr: string;
  descriptionEn: string;
}

export interface HomeContent {
  stats: Stat[];
  designStages: DesignStage[];
  differentiators: Differentiator[];
}

export interface PrintEstimateInput {
  /**
     * @minimum 1
     * @maximum 100000
     */
  weightGrams: number;
  materialId: string;
  /**
     * @minimum 1
     * @maximum 1000
     */
  quantity: number;
}

export interface PrintEstimate {
  materialId: string;
  quantity: number;
  weightGrams: number;
  estimatedCost: number;
  estimatedLeadDays: number;
  currency: string;
  noteAr: string;
  noteEn: string;
}

export interface InquiryInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  email: string;
  /** @maxLength 160 */
  company?: string;
  serviceSlug?: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  message: string;
}

export interface InquiryReceipt {
  reference: string;
  messageAr: string;
  messageEn: string;
}