# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /client/profile:
    get:
      operationId: getClientProfile
      tags: [client]
      summary: Get the authenticated client's profile
      security:
        - clerkSession: []
      responses:
        "200":
          description: Client profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProfile"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateClientProfile
      tags: [client]
      summary: Update the authenticated client's profile
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientProfileInput"
      responses:
        "200":
          description: Updated client profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProfile"
        "400":
          description: Invalid profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /client/overview:
    get:
      operationId: getClientOverview
      tags: [client]
      summary: Get the authenticated client's requests and activity
      security:
        - clerkSession: []
      responses:
        "200":
          description: Client overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientOverview"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /client/print-requests:
    post:
      operationId: createClientPrintRequest
      tags: [client]
      summary: Create an authenticated 3D printing request
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientPrintRequestInput"
      responses:
        "201":
          description: Persisted client request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientRequest"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Request service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /client/consultations:
    get:
      operationId: getClientConsultations
      tags: [client]
      summary: Get the authenticated client's consultation requests
      security:
        - clerkSession: []
      responses:
        "200":
          description: Consultation requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientConsultation"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Consultation service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createClientConsultation
      tags: [client]
      summary: Create an authenticated consultation or specialist request
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientConsultationInput"
      responses:
        "201":
          description: Persisted consultation request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientConsultation"
        "400":
          description: Invalid consultation request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Consultation service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    clerkSession:
      type: http
      scheme: bearer
      bearerFormat: Clerk session
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ClientProfile:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        email:
          type: string
        name:
          type: string
        company:
          type: string
      required: [userId, username, email, name, company]
    UpdateClientProfileInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        company:
          type: string
          maxLength: 160
      required: [name, company]
    ClientRequest:
      type: object
      properties:
        id:
          type: string
        reference:
          type: string
        kind:
          type: string
          enum: [print]
        projectName:
          type: string
        serviceSlug:
          type: string
        status:
          type: string
          enum: [submitted, reviewing, quoted, scheduled, completed, suspended]
        statusAr:
          type: string
        statusEn:
          type: string
        material:
          type: string
        finish:
          type: string
        quantity:
          type: integer
        timeline:
          type: string
        notes:
          type: string
        fileName:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, reference, kind, projectName, serviceSlug, status, statusAr, statusEn, material, finish, quantity, timeline, notes, createdAt]
    ClientOverview:
      type: object
      properties:
        profile:
          $ref: "#/components/schemas/ClientProfile"
        requests:
          type: array
          items:
            $ref: "#/components/schemas/ClientRequest"
      required: [profile, requests]
    CreateClientPrintRequestInput:
      type: object
      properties:
        projectName:
          type: string
          minLength: 2
          maxLength: 160
        material:
          type: string
          minLength: 1
          maxLength: 80
        quantity:
          type: integer
          minimum: 1
          maximum: 1000
        finish:
          type: string
          minLength: 1
          maxLength: 80
        timeline:
          type: string
          minLength: 1
          maxLength: 80
        notes:
          type: string
          maxLength: 3000
        fileName:
          type: string
          maxLength: 255
      required: [projectName, material, quantity, finish, timeline, notes]
    ClientConsultation:
      type: object
      properties:
        id:
          type: string
        reference:
          type: string
        kind:
          type: string
          enum: [consultation, specialist]
        status:
          type: string
          enum: [submitted, reviewing, contacted, completed, suspended]
        statusAr:
          type: string
        statusEn:
          type: string
        title:
          type: string
        details:
          type: string
        specialty:
          type: string
        providerType:
          type: string
          enum: [person, company, guide]
        preferredProvider:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, reference, kind, status, statusAr, statusEn, title, details, createdAt]
    ClientConsultationInput:
      type: object
      properties:
        kind:
          type: string
          enum: [consultation, specialist]
        title:
          type: string
          minLength: 2
          maxLength: 160
        details:
          type: string
          minLength: 10
          maxLength: 4000
        specialty:
          type: string
          maxLength: 120
        providerType:
          type: string
          enum: [person, company, guide]
        preferredProvider:
          type: string
          maxLength: 160
      required: [kind, title, details]
...